WIP, but compilable
