+++ raiden-derive/src/ops/query.rs
@@ -83,7 +83,7 @@ pub(crate) fn expand_query(
             //     self
             // }
 
-            pub fn key_condition(mut self, cond: impl ::raiden::Querium::QueriumBuilder<#key_condition_token_name>) -> Self {
+            pub fn key_condition(mut self, cond: impl ::raiden::querium::QueriumBuilder<#key_condition_token_name>) -> Self {
                 let (cond_str, attr_names, attr_values) = cond.build();
                 if !attr_values.is_empty() {
                     self.input.expression_attribute_values = Some(attr_values);
