+++ raiden/src/filter_expression/mod.rs
@@ -13,6 +13,17 @@ pub struct FilterExpression<T> {
     inner: Querium<T>,
 }
 
+impl<T> FilterExpression<T> {
+    fn new(inner: Querium<T>) -> Self {
+        Self { inner }
+    }
+
+    pub fn from_attr(attr: String) -> Self {
+        let inner = Querium::new(attr);
+        Self::new(inner)
+    }
+}
+
 #[derive(Debug, Clone)]
 pub struct FilterExpressionFilledOrWaitConjunction<T> {
     inner: QueriumFilledOrWaitConjunction<T>,
